refactor(pdf-overlay): extract types and session lifecycle
